Wire plashet as config:plashet verb
